# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DebateRoom API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rooms:
    get:
      operationId: listRooms
      tags: [rooms]
      summary: List debate rooms
      security:
        - bearerAuth: []
      parameters:
        - name: category
          in: query
          schema:
            $ref: "#/components/schemas/RoomCategory"
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: myRooms
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Rooms list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomsListResponse"
    post:
      operationId: createRoom
      tags: [rooms]
      summary: Create a debate room
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoomRequest"
      responses:
        "201":
          description: Room created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
  /rooms/similar:
    get:
      operationId: findSimilarRooms
      tags: [rooms]
      summary: Find rooms with similar titles in a category
      security:
        - bearerAuth: []
      parameters:
        - name: title
          in: query
          required: true
          schema:
            type: string
        - name: category
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/RoomCategory"
      responses:
        "200":
          description: Similar rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"
  /rooms/{roomId}:
    get:
      operationId: getRoom
      tags: [rooms]
      summary: Get room details
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Room details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateRoom
      tags: [rooms]
      summary: Update room (owner only)
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRoomRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
    delete:
      operationId: deleteRoom
      tags: [rooms]
      summary: Delete room (owner only)
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /rooms/{roomId}/join:
    post:
      operationId: joinRoom
      tags: [rooms]
      summary: Join a room
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Joined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /rooms/{roomId}/leave:
    post:
      operationId: leaveRoom
      tags: [rooms]
      summary: Leave a room
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Left
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /rooms/{roomId}/members:
    get:
      operationId: getRoomMembers
      tags: [rooms]
      summary: Get room members
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RoomMember"
  /rooms/{roomId}/members/{userId}:
    delete:
      operationId: removeRoomMember
      tags: [rooms]
      summary: Remove a member (owner only)
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Removed
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
    RoomCategory:
      type: string
      enum:
        - love
        - romance
        - money
        - business
        - religion
        - politics
        - sports
        - technology
        - lifestyle
        - other
...
    Profile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        username:
          type: string
        avatarKey:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - username
        - avatarKey
        - createdAt
...
    Room:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        category:
          $ref: "#/components/schemas/RoomCategory"
        ownerId:
          type: string
        createdAt:
          type: string
          format: date-time
        lastActivityAt:
          type: string
          format: date-time
        isArchived:
          type: boolean
        memberCount:
          type: integer
        lastMessage:
          type: string
          nullable: true
        isMember:
          type: boolean
        isOwner:
          type: boolean
      required:
        - id
        - title
        - category
        - ownerId
        - createdAt
        - lastActivityAt
        - isArchived
        - memberCount
        - isMember
        - isOwner
    RoomDetail:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        category:
          $ref: "#/components/schemas/RoomCategory"
        ownerId:
          type: string
        createdAt:
          type: string
          format: date-time
        lastActivityAt:
          type: string
          format: date-time
        isArchived:
          type: boolean
        memberCount:
          type: integer
        isMember:
          type: boolean
        isOwner:
          type: boolean
        owner:
          $ref: "#/components/schemas/Profile"
      required:
        - id
        - title
        - category
        - ownerId
        - createdAt
        - lastActivityAt
        - isArchived
        - memberCount
        - isMember
        - isOwner
        - owner
    RoomsListResponse:
      type: object
      properties:
        rooms:
          type: array
          items:
            $ref: "#/components/schemas/Room"
        total:
          type: integer
        hasMore:
          type: boolean
      required:
        - rooms
        - total
        - hasMore
    CreateRoomRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 100
        description:
          type: string
          maxLength: 300
        category:
          $ref: "#/components/schemas/RoomCategory"
      required:
        - title
        - category
    UpdateRoomRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 100
        description:
          type: string
          maxLength: 300
    RoomMember:
      type: object
      properties:
        userId:
          type: string
        username:
          type: string
        avatarKey:
          type: string
        role:
          type: string
          enum:
            - owner
            - member
        joinedAt:
          type: string
          format: date-time
      required:
        - userId
        - username
        - avatarKey
        - role
        - joinedAt
...